Changed the target function of Graph::optimize to user-defined cost function

+++ src/quartz/tasograph/tasograph.h
@@ -199,7 +199,7 @@ public:
            GateType target_rotation, std::string circuit_name = "",
            int timeout = 86400 /*1 day*/);
   std::shared_ptr<Graph> optimize(std::vector<GraphXfer *> xfers,
-                                  int gate_count_upper_bound,
+                                  double gate_count_upper_bound,
                                   std::string circuit_name, bool print_message,
                                   int timeout = 86400 /*1 day*/);
   void constant_and_rotation_elimination();
